Added $petr functionality
I probably put too much time into this but it was fun and now I know how to set up discord bots so now I'm *really* dangerous.

Added a bunch of modular functionality, images, etc. It should be fairly straightforward.

If you add more images, simply run build_petrs.py (as a script) to rebuild the csv file containing the names of the petr files and their weights. Then the bot should be good to go!

And feel free to change weights and things and add new weights, etc.